separate lib and cmd
